# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Fashion Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ProductVariant {
  id: number;
  name: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  colorCode?: string | null;
  price: string;
  inStock: boolean;
}

export interface Product {
  id: number;
  /** @nullable */
  printfulId?: number | null;
  name: string;
  category: string;
  basePrice: string;
  imageUrl: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  additionalImages?: string[];
}

export interface ProductDetail {
  id: number;
  /** @nullable */
  printfulId?: number | null;
  name: string;
  category: string;
  basePrice: string;
  imageUrl: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  additionalImages?: string[];
  /** @nullable */
  description?: string | null;
  variants: ProductVariant[];
  relatedProducts: Product[];
}

export interface CartItem {
  productId: number;
  variantId: number;
  name: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  price: string;
  quantity: number;
  imageUrl: string;
}

export interface CreateCheckoutSessionBody {
  items: CartItem[];
  successUrl: string;
  cancelUrl: string;
}

export interface CheckoutSessionResponse {
  url: string;
  sessionId: string;
}

export interface WebhookResponse {
  received: boolean;
}

export interface Order {
  id: number;
  stripeSessionId: string;
  /** @nullable */
  printfulOrderId?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  status: string;
  lineItems: CartItem[];
  createdAt: string;
}