# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Luxury Fashion Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/category/{slug}:
    get:
      operationId: getProductsByCategory
      tags: [products]
      summary: Get products by category
      description: Returns products filtered by category slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of products in category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get single product
      description: Returns a single product with all variants and real-time pricing
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        colorCode:
          type: ["string", "null"]
        price:
          type: string
        inStock:
          type: boolean
      required:
        - id
        - name
        - price
        - inStock
    Product:
      type: object
      properties:
        id:
          type: integer
        printfulId:
          type: ["integer", "null"]
        name:
          type: string
        category:
          type: string
        basePrice:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        additionalImages:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - category
        - basePrice
        - imageUrl
    ProductDetail:
      type: object
      properties:
        id:
          type: integer
        printfulId:
          type: ["integer", "null"]
        name:
          type: string
        category:
          type: string
        basePrice:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        additionalImages:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        relatedProducts:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required:
        - id
        - name
        - category
        - basePrice
        - imageUrl
        - variants
        - relatedProducts
...