# lib/api-zod/src/generated/types/productDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Fashion Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";
import type { ProductVariant } from "./productVariant";

export interface ProductDetail {
  id: number;
  /** @nullable */
  printfulId?: number | null;
  name: string;
  category: string;
  basePrice: string;
  imageUrl: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  additionalImages?: string[];
  /** @nullable */
  description?: string | null;
  variants: ProductVariant[];
  relatedProducts: Product[];
}