# lib/api-zod/src/generated/types/productVariant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Fashion Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductVariant {
  id: number;
  name: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  colorCode?: string | null;
  price: string;
  inStock: boolean;
}